requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.15
inflection==0.5.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
reportlab==4.4.10
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
